refactor: refactoring complex app
- Break Each View Into a Hierarchy of Components
- Adding search and main root
