+++ src/BookShelf.js
@@ -1,0 +1,56 @@
+import React, {Component} from 'react'
+import PropTypes from 'prop-types'
+import './App.css'
+import Book from './Book';
+
+class BookShelf extends Component {
+    state = {};
+    render() {
+        return (            
+        <div className="bookshelf">
+        <h2 className="bookshelf-title">{this.props.shelfTitle}</h2>
+        <div className="bookshelf-books">
+          <ol className="books-grid">
+          <li>
+                        <div className="book">
+                          <div className="book-top">
+                            <div className="book-cover" style={{ width: 128, height: 193, backgroundImage: 'url("http://books.google.com/books/content?id=uu1mC6zWNTwC&printsec=frontcover&img=1&zoom=1&imgtk=AFLRE73pGHfBNSsJG9Y8kRBpmLUft9O4BfItHioHolWNKOdLavw-SLcXADy3CPAfJ0_qMb18RmCa7Ds1cTdpM3dxAGJs8zfCfm8c6ggBIjzKT7XR5FIB53HHOhnsT7a0Cc-PpneWq9zX&source=gbs_api")' }}></div>
+                            <div className="book-shelf-changer">
+                              <select>
+                                <option value="move" disabled>Move to...</option>
+                                <option value="currentlyReading">Currently Reading</option>
+                                <option value="wantToRead">Want to Read</option>
+                                <option value="read">Read</option>
+                                <option value="none">None</option>
+                              </select>
+                            </div>
+                          </div>
+                          <div className="book-title">1776</div>
+                          <div className="book-authors">David McCullough</div>
+                        </div>
+                      </li>
+                      <li>
+                        <div className="book">
+                          <div className="book-top">
+                            <div className="book-cover" style={{ width: 128, height: 192, backgroundImage: 'url("http://books.google.com/books/content?id=wrOQLV6xB-wC&printsec=frontcover&img=1&zoom=1&imgtk=AFLRE72G3gA5A-Ka8XjOZGDFLAoUeMQBqZ9y-LCspZ2dzJTugcOcJ4C7FP0tDA8s1h9f480ISXuvYhA_ZpdvRArUL-mZyD4WW7CHyEqHYq9D3kGnrZCNiqxSRhry8TiFDCMWP61ujflB&source=gbs_api")' }}></div>
+                            <div className="book-shelf-changer">
+                              <select>
+                                <option value="move" disabled>Move to...</option>
+                                <option value="currentlyReading">Currently Reading</option>
+                                <option value="wantToRead">Want to Read</option>
+                                <option value="read">Read</option>
+                                <option value="none">None</option>
+                              </select>
+                            </div>
+                          </div>
+                          <div className="book-title">Harry Potter and the Sorcerer's Stone</div>
+                          <div className="book-authors">J.K. Rowling</div>
+                        </div>
+                      </li>
+                    </ol>
+        </div>
+      </div>);
+    };
+}
+
+export default BookShelf
